Split off x86 GC Info and unwinding code into separate sharable files
